# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WhereWeAre API",
    "version": "1.0.0",
    "description": "Comprehensive API for accessing cultural landmarks, user management, memories, and favorites in Chemnitz"
  },
  "servers": [
    {
      "url": "https://whereweare.vercel.app",
      "description": "Application server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token"
      }
    },
    "schemas": {
...
      "Favorite": {
        "type": "object",
        "properties": {
          "siteId": {
            "type": "string"
          },
          "siteName": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "coordinates": {
            "type": "object",
            "properties": {
              "lat": {
                "type": "number"
              },
              "lng": {
                "type": "number"
              }
            }
          },
          "addedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Activity": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "visit",
              "favorite",
              "memory",
              "search"
            ]
          },
          "siteId": {
            "type": "string"
          },
          "siteName": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "code": {
            "type": "string",
            "description": "Error code"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/user/activity": {
      "get": {
        "summary": "Get user's activity history",
        "description": "Retrieve the user's activity history including visits, favorites, memories, and searches. Activities are typically ordered by creation date.",
        "tags": [
          "Activities"
        ],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of user activities",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Activity"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new activity",
        "description": "Record a new user activity such as visiting a site, adding to favorites, creating a memory, or performing a search. All fields are required.",
        "tags": [
          "Activities"
        ],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "type",
                  "siteId",
                  "siteName",
                  "category"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "visit",
                      "favorite",
                      "memory",
                      "search"
                    ],
                    "description": "Type of activity"
                  },
                  "siteId": {
                    "type": "string",
                    "description": "Cultural site ID"
                  },
                  "siteName": {
                    "type": "string",
                    "description": "Name of the site"
                  },
                  "category": {
                    "type": "string",
                    "description": "Site category"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Activity added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "activityId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/favorites": {
      "get": {
        "summary": "Get user's favorite sites",
        "description": "Retrieve the user's favorite cultural sites with optional grouping. Use groupBy=category to group favorites by site category.",
        "tags": [
          "Favorites"
        ],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "groupBy",
            "schema": {
              "type": "string",
              "enum": [
                "category"
              ]
            },
            "description": "Group favorites by category"
          }
        ],
        "responses": {
          "200": {
            "description": "List of user's favorites",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Favorite"
                      }
                    },
                    {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/Favorite"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a site to favorites",
        "description": "Add a cultural site to the user's favorites list. siteId, siteName, and category are required.",
        "tags": [
          "Favorites"
        ],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "siteId",
                  "siteName",
                  "category"
                ],
                "properties": {
                  "siteId": {
                    "type": "string",
                    "description": "Cultural site ID"
                  },
                  "siteName": {
                    "type": "string",
                    "description": "Name of the site"
                  },
                  "category": {
                    "type": "string",
                    "description": "Site category"
                  },
                  "description": {
                    "type": "string",
                    "description": "Site description"
                  },
                  "coordinates": {
                    "type": "object",
                    "properties": {
                      "lat": {
                        "type": "number"
                      },
                      "lng": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Site added to favorites",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove a site from favorites",
        "description": "Remove a cultural site from the user's favorites list. siteId query parameter is required.",
        "tags": [
          "Favorites"
        ],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "siteId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Site ID to remove from favorites"
          }
        ],
        "responses": {
          "200": {
            "description": "Site removed from favorites",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Site ID is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}